Move DCAP attestation to `attestation_handler.rs`
